# Use faster Python script to amalgamate