@Cyan4973 2022-01-17
Thanks @cwoffenden :+1:  ,
I'm sure there will be quite some interest in this faster version :wink: