@cwoffenden 2022-01-19
This is done and tidied. I limited running the Python version to 3.8 (and higher) though it _may_ work on earlier versions (testing was only done on 3.8). The Python is as Python-like as I know how (I'm not a Python guy), has the type annotations and has been passed through `mypy` static type checking.

I changed the amalgamation scripts to use the `-x` flag to exclude `zstd_legacy.h` since both the decoder and full libraries build without `ZSTD_LEGACY_SUPPORT`, and this lets us remove 50kB from the resulting generated C (and if the macro is ever enabled an `#error` directive is inserted to show why it fails). I also tested when using the `-k` flag to not inline `zstd.h` to keep the library as separate header and implementation, and that also works well.

And it's fast, even on my 2013 Mac Pro:
```
carl@Betty single_file_libs % time ./combine.py -r ../../lib -x legacy/zstd_legacy.h -o zstddeclib.c zstddeclib-in.c
./combine.py -r ../../lib -x legacy/zstd_legacy.h -o zstddeclib.c   0.11s user 0.03s system 96% cpu 0.145 total
```
Consistently less than 0.2s vs nearly 3m previously, so over a thousand times faster.