@cwoffenden 2022-01-19
Added bonus: the Python script also works on Windows. In theory the shell script did, with the Git shell support, for example, but would take hours. It takes longer than on macOS on the same machine, completing in 0.6s:
```
Measure-Command {python3 combine.py -r ../../lib -x legacy/zstd_legacy.h -o zstddeclib.c zstddeclib-in.c}
Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 0
Milliseconds      : 594
```