@cwoffenden 2022-01-19
This has had further testing and edge cases fixed. I can now call it done.