@Cyan4973 2022-01-20
minor question :

It seems the new recipe needs to explicitly exclude `zstd_legacy.h`,
while that wasn't required before.
Any explanation why ?