@cwoffenden 2022-01-20
The input file always stripped out certain features, primarily to reduce the size when this was just the decompressor alone, e.g.:
```
#define ZSTD_LEGACY_SUPPORT 0
#define ZSTD_STRIP_ERROR_STRINGS
#define ZSTD_TRACE 0
```
In the original version of tool there wasn't the `-x` option, so all files were inlined, used or not.

I can add it back in, if you'd prefer? It removed 50k-ish from the generated source, which the compiler skips over because those parts are currently if'd out.