@Cyan4973 2022-01-20
I think it's a pretty good reason.
I believe it just needs to be documented, as otherwise people will just not understand why it's there, and may miss this optimization in some future update.