@cwoffenden 2022-01-20
I made the `#error` inserted into the generated code more details, and also added docs for the `ZSTD_LEGACY_SUPPORT` flag. I tidied the instructions and copyright year whilst I was in there.